fix for http://www.fossology.org/issues/1635 Add cast to variable print to remove warnings on mismatched sizes
git-svn-id: http://svn.code.sf.net/p/fossology/code/branches/fossology2.0/fossology@5651 a5643f74-2041-0410-a2b9-94ffef304a07

+++ src/nomos/agent/list.c
@@ -823,7 +823,7 @@ void listDump(list_t *l, int verbose) {
   if (verbose < 0) {
     printf("** %s (size %d, used %d, ix %d, sort %d desc %d) == %lu\n",
         l->name, l->size, l->used, l->ix, l->sorted, l->desc,
-        sizeof(item_t));
+        (unsigned long)sizeof(item_t));
     return;
   }
   if (verbose || max) {
